Control.Monad.Rescue.Class

Description

The MonadRescue class, meant for retrieving the success/failure branches

Documentation

class (Monad m, MonadRaise n) => MonadRescueFrom n m where

Pull a potential error out of the surrounding context NOTE that the target m may not even be aware of Raise/Rescue. It's an escape to the "normal" world

Methods

attempt :: n a -> m (Either (ErrorCase n) a)

Attempt some action, exposing the success and error branches

Examples

Expand

>>> :{
  goesBoom :: Int -> Rescue '[FooErr, BarErr] Int
  goesBoom x =
    if x > 50
      then return x
      else raise FooErr
 :}

>>> :{
  result :: Identity (Either (OpenUnion '[FooErr, BarErr]) Int)
  result = attempt $ goesBoom 42
:}

>>> result
Identity (Left (Identity FooErr))

Where Identity FooErr is the selection of the OpenUnion.