Control.Monad.Trans.Cleanup.Types

Description

The CleanupT transformer for adding async exceptions to a stack

Documentation

newtype CleanupT m a

Adds SomeException to an exception stack, and thus aware of async exceptions

Constructors

CleanupT

Fields

runCleanupT :: m a

type CleanupIO a = CleanupT IO a