Data.WorldPeace.Subset.Class

Documentation

class Subset err errs where

State that some type err is a subset of errs

In essence, this is used to avoid having to specially lift bare values to an OpenUnion. We can treat all types and OpenUnions the same.

Methods

include :: err -> errs