Control.Monad.Rescue.Class

Description

The MonadRescue class, meant for retrieving the success/failure branches

Documentation

class MonadRaise m => MonadRescue m where

Pull a potential error out of the surrounding context

Methods

attempt :: m a -> m (Either (OpenUnion (Errors m)) a)

Attempt some action, exposing the success and error branches

Examples

Expand

>>> :{
  goesBoom :: Int -> Rescue '[FooErr, BarErr] Int
  goesBoom x =
    if x > 50
      then return x
      else raise FooErr
 :}

>>> runRescue . attempt $ goesBoom 42
Right (Left (Identity FooErr))

Where Identity fooErr is the selection of the OpenUnion. In practice you would handle the OpenUnion like so:

>>> let handleErr = catchesOpenUnion (show, show)
>>> let x = attempt (goesBoom 42) >>= pure . either handleErr show
>>> runRescue x
Right "FooErr"